return raw data of chart render results
